Explicit cast before using thumbnailSize in string format.

+++ BoxBrowseSDK/BoxBrowseSDK/BOXThumbnailCache.m
@@ -77,7 +77,7 @@
 
 - (NSString *)cacheKeyForFile:(BOXFile *)file thumbnailSize:(BOXThumbnailSize)thumbnailSize
 {
-    NSString *key = [NSString stringWithFormat:@"%@_%lu", file.SHA1, thumbnailSize];
+    NSString *key = [NSString stringWithFormat:@"%@_%lu", file.SHA1, (unsigned long) thumbnailSize];
     return key;
 }
